Centralize project-bound execution env
