+++ brain/systems/runs/tool_catalog/handlers/files.py
@@ -2,10 +2,12 @@
 
 from __future__ import annotations
 
-import base64 as _base64
-from dataclasses import dataclass as _dataclass
-
 from brain.systems.runs.tool_catalog.handlers.common import *
+from brain.systems.runs.project_execution_env import (
+    annotate_project_execution_result as _annotate_project_execution_result,
+    prepare_project_execution_env as _prepare_project_execution_env,
+    redact_sensitive_output as _redact_sensitive_output,
+)
 
 
 def _resolve_path(path: str, working_dir: str | None = None) -> str:
@@ -38,115 +40,6 @@ _BLOCKED_PATTERNS = [
     _re.compile(r"chmod\s+-R\s+777\s+/"),
     _re.compile(r":\(\)\s*\{\s*:\|:&\s*\}"),  # fork bomb
 ]
-
-_GITHUB_TOKEN_ENV_NAMES = ("GH_TOKEN", "GITHUB_TOKEN")
-
-
-@_dataclass(frozen=True)
-class _ProjectExecutionEnv:
-    env: dict[str, str] | None
-    injected_env: list[str]
-    git_auth_hosts: list[str]
-    sensitive_values: list[str]
-
-
-def _github_token_from_project_env(project_env: dict[str, str]) -> str | None:
-    for env_name in _GITHUB_TOKEN_ENV_NAMES:
-        token = (project_env.get(env_name) or "").strip()
-        if token:
-            return token
-    return None
-
-
-def _append_git_config_env(env: dict[str, str], key: str, value: str) -> None:
-    try:
-        count = int(env.get("GIT_CONFIG_COUNT", "0") or "0")
-    except ValueError:
-        count = 0
-    env[f"GIT_CONFIG_KEY_{count}"] = key
-    env[f"GIT_CONFIG_VALUE_{count}"] = value
-    env["GIT_CONFIG_COUNT"] = str(count + 1)
-
-
-def _configure_project_bound_git_auth(
-    env: dict[str, str],
-    project_env: dict[str, str],
-) -> tuple[list[str], list[str]]:
-    """Let plain git use a project-bound GitHub token without persisting config."""
-    token = _github_token_from_project_env(project_env)
-    if not token:
-        return [], []
-
-    credential = f"x-access-token:{token}"
-    encoded = _base64.b64encode(credential.encode("utf-8")).decode("ascii")
-    auth_header = f"AUTHORIZATION: basic {encoded}"
-    _append_git_config_env(env, "http.https://github.com/.extraheader", auth_header)
-    env.setdefault("GIT_TERMINAL_PROMPT", "0")
-    env.setdefault("GCM_INTERACTIVE", "never")
-    return ["github.com"], [token, credential, encoded, auth_header]
-
-
-def _redact_sensitive_output(text: str, sensitive_values: list[str]) -> str:
-    redacted = text
-    values = sorted(
-        {value for value in sensitive_values if value},
-        key=len,
-        reverse=True,
-    )
-    for value in values:
-        redacted = redacted.replace(value, "[secret redacted]")
-    return redacted
-
-
-def _prepare_project_execution_env() -> _ProjectExecutionEnv:
-    project_env = _current_project_bound_env()
-    if not project_env:
-        return _ProjectExecutionEnv(
-            env=None,
-            injected_env=[],
-            git_auth_hosts=[],
-            sensitive_values=[],
-        )
-
-    run_env = os.environ.copy()
-    run_env.update(project_env)
-    git_auth_hosts, git_sensitive_values = _configure_project_bound_git_auth(run_env, project_env)
-    return _ProjectExecutionEnv(
-        env=run_env,
-        injected_env=sorted(project_env),
-        git_auth_hosts=git_auth_hosts,
-        sensitive_values=list(project_env.values()) + git_sensitive_values,
-    )
-
-
-def _annotate_project_execution_result(result: dict, project_execution: _ProjectExecutionEnv) -> None:
-    if project_execution.injected_env:
-        result["injected_env"] = project_execution.injected_env
-    if project_execution.git_auth_hosts:
-        result["git_auth_configured"] = project_execution.git_auth_hosts
-
-
-def _current_project_bound_env() -> dict[str, str]:
-    """Resolve project-bound vault tokens for command env injection."""
-    user_id = getattr(_agent_context, "user_id", None)
-    if not user_id:
-        return {}
-    project_context = _current_project_token_context()
-    if not project_context.get("project_slug") and project_context.get("target_registry_id") is None:
-        return {}
-    try:
-        from brain.systems.vault import resolve_project_bound_env_tokens
-
-        return resolve_project_bound_env_tokens(
-            user_id=user_id,
-            org_id=getattr(_agent_context, "org_id", None),
-            project_slug=project_context.get("project_slug"),
-            project_slugs=project_context.get("project_slugs"),
-            target_registry_id=project_context.get("target_registry_id"),
-        )
-    except Exception:
-        logger.debug("project_bound_vault_env_resolution_failed", exc_info=True)
-        return {}
 
 
 def _handle_exec_command(
